Fix ban commands. Move level command to gm.
